requirements.txt
django==4.2.4
djangorestframework==3.14.0
scikit-learn==1.2.2
joblib==1.3.2
pandas==1.3.3
